+++ src/mock/v1.2/channelService-1_2.ts
@@ -1,6 +1,6 @@
 import { Channel, DesktopAgent } from "fdc3_1_2";
 import constants from "../../constants";
-import { ChannelsAppConfig } from "../../test/common/channel-control";
+import { ChannelsAppConfig } from "../../test/common/control/channel-control";
 import { AppControlContext } from "../../context-types";
 import { commands, channelType } from "../constants";
 declare let fdc3: DesktopAgent;
