+++ src/mock/v2.0/channelService-2_0.ts
@@ -1,6 +1,6 @@
 import { Channel, DesktopAgent } from "fdc3_2_0";
 import constants from "../../constants";
-import { ChannelsAppConfig } from "../../test/common/channel-control";
+import { ChannelsAppConfig } from "../../test/common/control/channel-control";
 import { AppControlContext } from "../../context-types";
 import { commands, channelType } from "../constants";
 declare let fdc3: DesktopAgent;
