+++ src/test/common/fdc3.user-channels.ts
@@ -1,8 +1,8 @@
 import { assert, expect } from "chai";
 import constants from "../../constants";
 import { failOnTimeout, wait, wrapPromise } from "../../utils";
-import { JOIN_AND_BROADCAST, JOIN_AND_BROADCAST_TWICE } from "../common/channel-control";
-import { ChannelControl } from "./channel-control";
+import { JOIN_AND_BROADCAST, JOIN_AND_BROADCAST_TWICE } from "./control/channel-control";
+import { ChannelControl } from "./control/channel-control";
 
 export function createUserChannelTests(cc: ChannelControl<any, any, any>, documentation: string, prefix: string): Mocha.Suite {
   const channelName = prefix === "" ? "System channels" : "User channels";
