+++ src/test/v1.2/advanced/channels.ts
@@ -1,5 +1,5 @@
 import { APIDocumentation1_2 } from "../apiDocuments-1.2";
-import { ChannelControl1_2 } from "./channels-support-1.2";
+import { ChannelControl1_2 } from "../support/channels-support-1.2";
 import { createAppChannelTests } from "../../common/fdc3.app-channels";
 import { createUserChannelTests } from "../../common/fdc3.user-channels";
