+++ src/test/v2.0/advanced/fdc3.raiseIntent.ts
@@ -1,7 +1,7 @@
 import { ChannelError, PrivateChannel } from "fdc3_2_0";
 import { assert, expect } from "chai";
 import { wait } from "../../../utils";
-import { RaiseIntentControl2_0, IntentResultType, IntentApp } from "./intent-support-2.0";
+import { RaiseIntentControl2_0, IntentResultType, IntentApp } from "../support/intent-support-2.0";
 import constants from "../../../constants";
 import { closeMockAppWindow } from "../fdc3-2_0-utils";
