+++ src/test/v2.0/support/open-support-2.0.ts
@@ -4,7 +4,7 @@ import constants from "../../../constants";
 import { ContextSender } from "../../../mock/v2.0/general";
 import { sleep } from "../../../utils";
 import { AppControlContext } from "../../../context-types";
-import { OpenControl } from "../../common/open-control";
+import { OpenControl } from "../../common/control/open-control";
 import { APIDocumentation2_0 } from "../apiDocuments-2.0";
 import { closeMockAppWindow } from "../fdc3-2_0-utils";
